+++ tests/ha/single_pod_restart/test_data_pod_restart.py
@@ -72,10 +72,9 @@ class TestDataPodRestart:
         cls.hlth_master_list = []
         cls.node_worker_list = []
         cls.ha_obj = HAK8s()
-        cls.random_time = cls.s3_clean = cls.test_prefix = cls.test_prefix_deg = None
+        cls.s3_clean = cls.test_prefix = cls.test_prefix_deg = cls.set_name = None
         cls.s3acc_name = cls.s3acc_email = cls.bucket_name = cls.object_name = cls.node_name = None
         cls.restore_pod = cls.deployment_backup = cls.deployment_name = cls.restore_method = None
-        cls.multipart_obj_path = cls.set_name = None
         cls.mgnt_ops = ManagementOPs()
         cls.system_random = secrets.SystemRandom()
 
@@ -97,7 +96,6 @@ class TestDataPodRestart:
 
         cls.rest_obj = S3AccountOperations()
         cls.s3_mp_test_obj = S3MultipartTestLib(endpoint_url=S3_CFG["s3_url"])
-        cls.test_file = "ha-mp_obj"
         cls.test_dir_path = os.path.join(TEST_DATA_FOLDER, "HATestMultipartUpload")
 
     def setup_method(self):
@@ -105,18 +103,16 @@ class TestDataPodRestart:
         This function will be invoked prior to each test case.
         """
         LOGGER.info("STARTED: Setup Operations")
-        self.random_time = int(time.time())
         self.restore_pod = False
         self.s3_clean = dict()
         self.s3acc_name = f"ha_s3acc_{int(perf_counter_ns())}"
         self.s3acc_email = f"{self.s3acc_name}@seagate.com"
-        self.bucket_name = f"ha-mp-bkt-{self.random_time}"
-        self.object_name = f"ha-mp-obj-{self.random_time}"
+        self.bucket_name = f"ha-mp-bkt-{int(perf_counter_ns())}"
+        self.object_name = f"ha-mp-obj-{int(perf_counter_ns())}"
         self.extra_files = list()
         if not os.path.exists(self.test_dir_path):
             resp = system_utils.make_dirs(self.test_dir_path)
             LOGGER.info("Created path: %s", resp)
-        self.multipart_obj_path = os.path.join(self.test_dir_path, self.test_file)
         LOGGER.info("Precondition: Verify cluster is up and running and all pods are online.")
         resp = self.ha_obj.check_cluster_status(self.node_master_list[0])
         assert_utils.assert_true(resp[0], resp[1])
@@ -537,459 +533,6 @@ class TestDataPodRestart:
 
     @pytest.mark.ha
     @pytest.mark.lc
-    @pytest.mark.skip(reason="Multipart upload is not supported with DTMInt0")
-    @pytest.mark.tags("TEST-34080")
-    def test_mpu_after_pod_restart(self):
-        """
-        This test tests multipart upload after data pod restart.
-        """
-        LOGGER.info("STARTED: Test to verify multipart upload after data pod restart.")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = HA_CFG["5gb_mpu_data"]["total_parts"]
-        download_path_1 = os.path.join(self.test_dir_path, self.test_file + "_download_1")
-        download_path_2 = os.path.join(self.test_dir_path, self.test_file + "_download_2")
-        download_path = os.path.join(self.test_dir_path, self.test_file + "_download")
-        LOGGER.info("Step 1: Create and list buckets. Perform multipart upload for size %s MB in "
-                    "total %s parts.", file_size, total_parts)
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.debug("Response: %s", resp)
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=self.bucket_name,
-                                                         object_name=self.object_name,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", self.bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum_1 = str(resp[2])
-        LOGGER.info("Step 1: Successfully performed multipart upload for size %s MB in total %s "
-                    "parts.", file_size, total_parts)
-        LOGGER.info("Step 2: Shutdown random data pod with replica method and "
-                    "verify cluster & remaining pods status")
-        num_replica = self.num_replica - 1
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            delete_pod=[self.delete_pod], num_replica=num_replica)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        pod_name = list(resp[1].keys())[0]
-        self.set_name = resp[1][pod_name]['deployment_name']
-        self.restore_method = resp[1][pod_name]['method']
-        pod_name = list(resp[1].keys())[0]
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.", pod_name)
-        self.restore_pod = True
-        LOGGER.info("Step 3: Download the uploaded object %s in healthy cluster & verify checksum",
-                    self.object_name)
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, self.bucket_name, self.object_name,
-                                                  download_path_1, upload_checksum_1)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 3: Successfully downloaded the object %s & verified the checksum",
-                    self.object_name)
-        object_name_1 = f"ha-mp-obj-{int(perf_counter_ns())}"
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bucket_name_1 = f"ha-mp-bkt-{int(perf_counter_ns())}"
-        else:
-            bucket_name_1 = self.bucket_name
-        LOGGER.info("Step 4: Perform multipart upload for size %s MB in total %s parts.",
-                    file_size, total_parts)
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=bucket_name_1,
-                                                         object_name=object_name_1,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(bucket_name_1, object_name_1)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", bucket_name_1, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum_2 = str(resp[2])
-        LOGGER.info("Step 4: Successfully performed multipart upload for  size %s MB in "
-                    "total %s parts.", file_size, total_parts)
-        LOGGER.info("Step 5: Start pod again and check cluster status")
-        resp = self.ha_obj.restore_pod(pod_obj=self.node_master_list[0],
-                                       restore_method=self.restore_method,
-                                       restore_params={"deployment_name": self.deployment_name,
-                                                       "deployment_backup":
-                                                           self.deployment_backup,
-                                                       "num_replica": self.num_replica,
-                                                       "set_name": self.set_name},
-                                       clstr_status=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to restore pod by {self.restore_method} way")
-        self.restore_pod = False
-        LOGGER.info("Step 5: Successfully started pod again and cluster is online")
-
-        LOGGER.info("Step 6.1: Download the uploaded object %s in healthy cluster & verify "
-                    "checksum", self.object_name)
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, self.bucket_name, self.object_name,
-                                                  download_path_1, upload_checksum_1)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 6.1: Successfully downloaded the object %s & verified the checksum",
-                    self.object_name)
-
-        LOGGER.info("Step 6.2: Download the uploaded object %s in degraded cluster & verify "
-                    "checksum", object_name_1)
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, bucket_name_1, object_name_1,
-                                                  download_path_2, upload_checksum_2)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 6.2: Successfully downloaded the object %s & verified the checksum",
-                    object_name_1)
-        LOGGER.info("Step 7: Perform multipart upload for size %s MB in total %s parts.",
-                    file_size, total_parts)
-        test_object = f"ha-mp-obj-{int(perf_counter_ns())}"
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            test_bucket = f"ha-mp-bkt-{int(perf_counter_ns())}"
-        else:
-            test_bucket = self.bucket_name
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=test_bucket,
-                                                         object_name=test_object,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(test_bucket, test_object)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", test_bucket, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum = str(resp[2])
-        LOGGER.info("Step 7: Successfully performed multipart upload for  size %s MB in "
-                    "total %s parts.", file_size, total_parts)
-        LOGGER.info("Step 8: Download the uploaded object %s & verify checksum", test_object)
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, test_bucket, test_object,
-                                                  download_path, upload_checksum)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 8: Successfully downloaded the object %s & verified the checksum",
-                    test_object)
-        self.extra_files.extend((self.multipart_obj_path, download_path, download_path_1,
-                                 download_path_2))
-        LOGGER.info("COMPLETED: Test to verify multipart upload after data pod restart.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.skip(reason="Multipart upload is not supported with DTMInt0")
-    @pytest.mark.tags("TEST-34082")
-    def test_partial_mpu_after_pod_restart(self):
-        """
-        This test tests partial multipart upload after data pod restart
-        """
-        LOGGER.info("STARTED: Test to verify partial multipart upload after data pod restart.")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = HA_CFG["5gb_mpu_data"]["total_parts"]
-        part_numbers = self.system_random.sample(list(range(1, total_parts + 1)), total_parts // 4)
-        download_path = os.path.join(self.test_dir_path, self.test_file + "_download")
-        if os.path.exists(self.multipart_obj_path):
-            os.remove(self.multipart_obj_path)
-        system_utils.create_file(self.multipart_obj_path, file_size)
-        LOGGER.info("Calculating checksum of file %s", self.multipart_obj_path)
-        upload_checksum = self.ha_obj.cal_compare_checksum(file_list=[self.multipart_obj_path],
-                                                           compare=False)[0]
-        LOGGER.info("Step 1: Start multipart upload for 5GB object in multiple parts and complete "
-                    "partially for %s part out of %s", part_numbers, total_parts)
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        s3_mp_test_obj = S3MultipartTestLib(access_key=access_key, secret_key=secret_key,
-                                            endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                    bucket_name=self.bucket_name,
-                                                    object_name=self.object_name,
-                                                    part_numbers=part_numbers,
-                                                    multipart_obj_size=file_size,
-                                                    total_parts=total_parts,
-                                                    multipart_obj_path=self.multipart_obj_path)
-        mpu_id = resp[1]
-        object_path = resp[2]
-        parts_etag1 = resp[3]
-        assert_utils.assert_true(resp[0], f"Failed to upload parts. Response: {resp}")
-        LOGGER.info("Step 1: Successfully completed partial multipart upload for %s part out of "
-                    "%s", part_numbers, total_parts)
-        LOGGER.info("Step 2: Listing parts of partial multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        for part_n in res[1]["Parts"]:
-            assert_utils.assert_list_item(part_numbers, part_n["PartNumber"])
-        LOGGER.info("Step 2: Listed parts of partial multipart upload: %s", res[1])
-        LOGGER.info("Step 3: Shutdown random data pod with replica method and "
-                    "verify cluster & remaining pods status")
-        num_replica = self.num_replica - 1
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            delete_pod=[self.delete_pod], num_replica=num_replica)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        pod_name = list(resp[1].keys())[0]
-        self.set_name = resp[1][pod_name]['deployment_name']
-        self.restore_method = resp[1][pod_name]['method']
-        pod_name = list(resp[1].keys())[0]
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 3: Successfully shutdown data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.", pod_name)
-        self.restore_pod = True
-        remaining_parts = list(filter(lambda i: i not in part_numbers,
-                                      list(range(1, total_parts + 1))))
-        parts_half = self.system_random.sample(remaining_parts, total_parts // 2)
-        part_numbers.extend(parts_half)
-        LOGGER.info("Step 4: Start multipart upload for 5GB object in multiple parts and complete "
-                    "partially for %s part out of %s", parts_half, total_parts)
-        resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                    bucket_name=self.bucket_name,
-                                                    object_name=self.object_name,
-                                                    part_numbers=parts_half,
-                                                    remaining_upload=True, mpu_id=mpu_id,
-                                                    multipart_obj_size=file_size,
-                                                    total_parts=total_parts,
-                                                    multipart_obj_path=object_path)
-        object_path = resp[2]
-        parts_etag2 = resp[3]
-        assert_utils.assert_true(resp[0], f"Failed to upload parts. Response: {resp}")
-        LOGGER.info("Step 4: Successfully completed partial multipart upload for %s part out of "
-                    "%s", parts_half, total_parts)
-        LOGGER.info("Step 5: Listing parts of partial multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        for part_n in res[1]["Parts"]:
-            assert_utils.assert_list_item(part_numbers, part_n["PartNumber"])
-        LOGGER.info("Step 5: Listed parts of partial multipart upload: %s", res[1])
-        LOGGER.info("Step 6: Start pod again and check cluster status")
-        resp = self.ha_obj.restore_pod(pod_obj=self.node_master_list[0],
-                                       restore_method=self.restore_method,
-                                       restore_params={"deployment_name": self.deployment_name,
-                                                       "deployment_backup":
-                                                           self.deployment_backup,
-                                                       "num_replica": self.num_replica,
-                                                       "set_name": self.set_name},
-                                       clstr_status=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to restore pod by {self.restore_method} way")
-        self.restore_pod = False
-        LOGGER.info("Step 6: Successfully started pod again and cluster is online")
-        remaining_parts = list(filter(lambda i: i not in part_numbers,
-                                      list(range(1, total_parts + 1))))
-        LOGGER.info("Step 7: Upload remaining %s parts out of %s", remaining_parts, total_parts)
-        resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                    bucket_name=self.bucket_name,
-                                                    object_name=self.object_name,
-                                                    part_numbers=remaining_parts,
-                                                    remaining_upload=True, mpu_id=mpu_id,
-                                                    multipart_obj_size=file_size,
-                                                    total_parts=total_parts,
-                                                    multipart_obj_path=object_path)
-        assert_utils.assert_true(resp[0], f"Failed to upload parts {resp[1]}")
-        parts_etag3 = resp[3]
-        LOGGER.info("Step 7: Successfully uploaded remaining %s parts out of %s",
-                    remaining_parts, total_parts)
-        etag_list = parts_etag1 + parts_etag2 + parts_etag3
-        parts_etag = sorted(etag_list, key=lambda d: d['PartNumber'])
-        LOGGER.info("Step 8: Listing parts of multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        assert_utils.assert_equal(len(res[1]["Parts"]), total_parts)
-        LOGGER.info("Step 8: Listed parts of multipart upload, Count: %s", len(res[1]["Parts"]))
-        LOGGER.info("Step 9: Completing multipart upload & check upload size is %s", file_size *
-                    const.Sizes.MB)
-        res = s3_mp_test_obj.complete_multipart_upload(mpu_id, parts_etag, self.bucket_name,
-                                                       self.object_name)
-        assert_utils.assert_true(res[0], res)
-        res = s3_test_obj.object_list(self.bucket_name)
-        if self.object_name not in res[1]:
-            assert_utils.assert_true(False, res)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", self.bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        LOGGER.info("Step 9: Multipart upload completed and verified upload size is %s",
-                    file_size * const.Sizes.MB)
-        LOGGER.info("Step 10: Download the uploaded object and verify checksum")
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, self.bucket_name, self.object_name,
-                                                  download_path, upload_checksum)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 10: Successfully downloaded the object and verified the checksum")
-        self.extra_files.extend((self.multipart_obj_path, download_path))
-        LOGGER.info("COMPLETED: Test to verify partial multipart upload after data pod restart.")
-
-    # pylint: disable=too-many-branches
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-34083")
-    def test_copy_obj_after_pod_restart(self):
-        """
-        This test tests copy object after data pod restart
-        """
-        LOGGER.info("STARTED: Test to verify copy object after data pod restart.")
-        bkt_cnt = HA_CFG["copy_obj_data"]["bkt_cnt"]
-        bkt_obj_dict = dict()
-        t_t = int(perf_counter_ns())
-        for cnt in range(bkt_cnt):
-            bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-        event = threading.Event()
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        LOGGER.info("Step 1: Create and list buckets. Upload object to %s & copy object from the"
-                    " same bucket to other buckets and verify copy object etags", self.bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=self.bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  file_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        put_etag = resp[1]
-        LOGGER.info("Step 1: Successfully created multiple buckets and uploaded object to %s "
-                    "and copied to other buckets and verified copy object etags", self.bucket_name)
-        LOGGER.info("Step 2: Shutdown random data pod with replica method and "
-                    "verify cluster & remaining pods status")
-        num_replica = self.num_replica - 1
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            delete_pod=[self.delete_pod], num_replica=num_replica)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        pod_name = list(resp[1].keys())[0]
-        self.set_name = resp[1][pod_name]['deployment_name']
-        self.restore_method = resp[1][pod_name]['method']
-        pod_name = list(resp[1].keys())[0]
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.", pod_name)
-        self.restore_pod = True
-        LOGGER.info("Step 3: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 3: Downloaded copied objects & verify etags.")
-        bkt_obj_dict1 = bkt_obj_dict.copy()
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{self.random_time}-1"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-        LOGGER.info("Step 4: Copy object from the %s bucket to other buckets and verify copy object"
-                    " etags", bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  file_path=self.multipart_obj_path, bkt_op=bkt_op)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        put_etag = resp[1]
-        LOGGER.info("Step 4: Successfully created multiple buckets and copied object from the %s "
-                    "bucket to other buckets and verified copy object etags", self.bucket_name)
-        LOGGER.info("Step 5: Start pod again and check cluster status")
-        resp = self.ha_obj.restore_pod(pod_obj=self.node_master_list[0],
-                                       restore_method=self.restore_method,
-                                       restore_params={"deployment_name": self.deployment_name,
-                                                       "deployment_backup":
-                                                           self.deployment_backup,
-                                                       "num_replica": self.num_replica,
-                                                       "set_name": self.set_name},
-                                       clstr_status=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to restore pod by {self.restore_method} way")
-        self.restore_pod = False
-        LOGGER.info("Step 5: Successfully started pod again and cluster is online")
-        LOGGER.info("Step 6: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        for bkt, obj in bkt_obj_dict1.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 6: Downloaded copied objects & verify etags.")
-
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{self.random_time}-2"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-        LOGGER.info("Step 7: Perform copy object from already created/uploaded %s bucket to other "
-                    "buckets verify copy object etags", bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict, put_etag=put_etag,
-                                                  bkt_op=bkt_op)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        LOGGER.info("Step 7: Performed copy object from already created/uploaded %s bucket to other"
-                    " buckets verified copy object etags", self.bucket_name)
-        LOGGER.info("Step 8: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 8: Downloaded copied objects & verify etags.")
-        LOGGER.info("COMPLETED: Test to verify copy object after data pod restart.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
     @pytest.mark.tags("TEST-34073")
     def test_reads_during_pod_restart(self):
         """
@@ -1048,6 +591,7 @@ class TestDataPodRestart:
                                   args=(event,), kwargs=args)
         thread.daemon = True  # Daemonize thread
         thread.start()
+        thread1 = None
         if CMN_CFG["dtm0_disabled"]:
             self.test_prefix = 'test-34073-deg'
             args = {'s3userinfo': list(users.values())[0], 'log_prefix': self.test_prefix,
@@ -1105,402 +649,6 @@ class TestDataPodRestart:
         assert_utils.assert_true(resp[0], resp[1])
         LOGGER.info("Step 8: Successfully ran IOs")
         LOGGER.info("ENDED: Test to verify continuous READs during data pod restart.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.skip(reason="Multipart upload is not supported with DTMInt0")
-    @pytest.mark.tags("TEST-34081")
-    def test_mpu_during_pod_restart(self):
-        """
-        This test tests multipart upload during data pod restart
-        """
-        LOGGER.info("STARTED: Test to verify multipart upload during data pod restart")
-
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = HA_CFG["5gb_mpu_data"]["total_parts"]
-        part_numbers = list(range(1, total_parts + 1))
-        random.shuffle(part_numbers)
-        output = Queue()
-        parts_etag = list()
-        download_path = os.path.join(self.test_dir_path, self.test_file + "_download")
-        download_path1 = os.path.join(self.test_dir_path, self.test_file + "_download1")
-        download_path2 = os.path.join(self.test_dir_path, self.test_file + "_download2")
-        event = threading.Event()  # Event to be used to send intimation of pod restart
-
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        s3_mp_test_obj = S3MultipartTestLib(access_key=access_key, secret_key=secret_key,
-                                            endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user")
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        LOGGER.info("Step 1: Perform multipart upload for size %s MB in total %s parts.",
-                    file_size, total_parts)
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=self.bucket_name,
-                                                         object_name=self.object_name,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", self.bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum1 = str(resp[2])
-        LOGGER.info("Step 1: Successfully performed multipart upload for  size %s MB in "
-                    "total %s parts.", file_size, total_parts)
-
-        LOGGER.info("Step 2: Shutdown random data pod with replica method and "
-                    "verify cluster & remaining pods status")
-        num_replica = self.num_replica - 1
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            delete_pod=[self.delete_pod], num_replica=num_replica)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        pod_name = list(resp[1].keys())[0]
-        self.set_name = resp[1][pod_name]['deployment_name']
-        self.restore_method = resp[1][pod_name]['method']
-        pod_name = list(resp[1].keys())[0]
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.", pod_name)
-        self.restore_pod = True
-
-        LOGGER.info("Step 3: Download the uploaded object in healthy cluster and verify checksum")
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, self.bucket_name,
-                                                  self.object_name, download_path, upload_checksum1)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 3: Successfully downloaded the object and verified the checksum")
-
-        object_name_1 = f"ha-mp-obj-{int(perf_counter_ns())}"
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create new bucket")
-            bucket_name_1 = f"ha-mp-bkt-{int(perf_counter_ns())}"
-        else:
-            bucket_name_1 = self.bucket_name
-
-        LOGGER.info("Step 4: Start multipart upload of %s MB object in background", file_size)
-        args = {'s3_data': self.s3_clean, 'bucket_name': bucket_name_1,
-                'object_name': object_name_1, 'file_size': file_size, 'total_parts': total_parts,
-                'multipart_obj_path': self.multipart_obj_path, 'part_numbers': part_numbers,
-                'parts_etag': parts_etag, 'output': output}
-        thread = threading.Thread(target=self.ha_obj.start_random_mpu, args=(event,), kwargs=args)
-        thread.daemon = True  # Daemonize thread
-        thread.start()
-        LOGGER.info("Step 4: Started multipart upload of %s MB object in background", file_size)
-
-        time.sleep(HA_CFG["common_params"]["60sec_delay"])
-
-        LOGGER.info("Step 5: Starting pod again and checking cluster status")
-        event.set()
-        resp = self.ha_obj.restore_pod(pod_obj=self.node_master_list[0],
-                                       restore_method=self.restore_method,
-                                       restore_params={"deployment_name": self.deployment_name,
-                                                       "deployment_backup":
-                                                           self.deployment_backup,
-                                                       "num_replica": self.num_replica,
-                                                       "set_name": self.set_name},
-                                       clstr_status=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to restore pod by {self.restore_method} way")
-        LOGGER.info("Step 5: Successfully started the pod and cluster is online")
-        self.restore_pod = False
-
-        LOGGER.info("Step 6: Checking responses from background process")
-        thread.join()
-        responses = tuple()
-        while len(responses) < 4:
-            responses = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-
-        if not responses:
-            assert_utils.assert_true(False, "Background process failed to do multipart upload")
-
-        exp_failed_parts = responses[0]
-        failed_parts = responses[1]
-        parts_etag = responses[2]
-        mpu_id = responses[3]
-        LOGGER.debug("Responses received from background process:\nexp_failed_parts: "
-                     "%s\nfailed_parts: %s\nparts_etag: %s\nmpu_id: %s", exp_failed_parts,
-                     failed_parts, parts_etag, mpu_id)
-        if len(exp_failed_parts) == 0 and len(failed_parts) == 0:
-            LOGGER.info("All the parts are uploaded successfully")
-        elif exp_failed_parts or failed_parts:
-            assert_utils.assert_true(False, "Failed to upload parts when cluster was in good "
-                                            f"state. Failed parts: {failed_parts} and "
-                                            f"{exp_failed_parts}")
-        LOGGER.info("Step 6: Successfully checked background process responses")
-
-        parts_etag = sorted(parts_etag, key=lambda d: d['PartNumber'])
-        LOGGER.info("Calculating checksum of file %s", self.multipart_obj_path)
-        upload_checksum2 = self.ha_obj.cal_compare_checksum(file_list=[self.multipart_obj_path],
-                                                            compare=False)[0]
-        LOGGER.info("Successfully uploaded all the parts of multipart upload.")
-
-        LOGGER.info("Step 7: Listing parts of multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, bucket_name_1, object_name_1)
-        assert_utils.assert_true(res[0], res)
-        assert_utils.assert_equal(len(res[1]["Parts"]), total_parts)
-        LOGGER.info("Step 7: Listed parts of multipart upload. Count: %s", len(res[1]["Parts"]))
-
-        LOGGER.info("Step 8: Completing multipart upload")
-        res = s3_mp_test_obj.complete_multipart_upload(mpu_id, parts_etag, bucket_name_1,
-                                                       object_name_1)
-        assert_utils.assert_true(res[0], res)
-        res = s3_test_obj.object_list(bucket_name_1)
-        assert_utils.assert_in(object_name_1, res[1], res)
-        LOGGER.info("Step 8: Multipart upload completed")
-
-        LOGGER.info("Step 9.1: Download the uploaded objects in healthy cluster and verify "
-                    "checksum")
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, self.bucket_name, self.object_name,
-                                                  download_path, upload_checksum1)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 9.2: Download the uploaded objects in step 4 and verify checksum")
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, bucket_name_1, object_name_1,
-                                                  download_path1, upload_checksum2)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 9: Successfully downloaded the objects and verified the checksum")
-
-        object_name_2 = f"ha-mp-obj-{int(perf_counter_ns())}"
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create new bucket")
-            bucket_name_2 = f"ha-mp-bkt-{int(perf_counter_ns())}"
-        else:
-            bucket_name_2 = self.bucket_name
-        LOGGER.info("Step 10: Again perform multipart upload for size %s MB in total %s parts and "
-                    "download the object and verify checksum", file_size, total_parts)
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=bucket_name_2,
-                                                         object_name=object_name_2,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(bucket_name_2, object_name_2)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", bucket_name_2, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum = str(resp[2])
-        LOGGER.info("Successfully performed multipart upload for size %s MB in total %s parts.",
-                    file_size, total_parts)
-        resp = self.ha_obj.dnld_obj_verify_chcksm(s3_test_obj, bucket_name_2, object_name_2,
-                                                  download_path2, upload_checksum)
-        assert_utils.assert_true(resp[0], resp[1])
-        LOGGER.info("Step 10: Successfully performed multipart upload and downloaded the object "
-                    "and verified the checksum")
-        self.extra_files.extend((self.multipart_obj_path, download_path, download_path1,
-                                 download_path2))
-        LOGGER.info("ENDED: Test to verify multipart upload during data pod restart")
-
-    # pylint: disable=too-many-branches
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-34084")
-    def test_copy_object_during_pod_restart(self):
-        """
-        This test tests copy object during data pod restart
-        """
-        LOGGER.info("STARTED: Test to verify copy object during data pod restart")
-
-        bkt_cnt = HA_CFG["copy_obj_data"]["bkt_cnt"]
-        bkt_obj_dict = dict()
-        t_t = int(perf_counter_ns())
-        for cnt in range(bkt_cnt):
-            bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-        output = Queue()
-        event = threading.Event()
-
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        LOGGER.info("Step 1: Create and list buckets. Upload object to %s & copy object from the"
-                    " same bucket to other buckets and verify copy object etags", self.bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=self.bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  file_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        put_etag = resp[1]
-        LOGGER.info("Step 1: Successfully created multiple buckets and uploaded object to %s "
-                    "and copied to other buckets and verified copy object etags", self.bucket_name)
-
-        LOGGER.info("Step 2: Shutdown random data pod with replica method and "
-                    "verify cluster & remaining pods status")
-        num_replica = self.num_replica - 1
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            delete_pod=[self.delete_pod], num_replica=num_replica)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        pod_name = list(resp[1].keys())[0]
-        self.set_name = resp[1][pod_name]['deployment_name']
-        self.restore_method = resp[1][pod_name]['method']
-        pod_name = list(resp[1].keys())[0]
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.", pod_name)
-        self.restore_pod = True
-
-        LOGGER.info("Step 3: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 3: Downloaded copied objects & verify etags.")
-
-        bkt_obj_dict1 = bkt_obj_dict.copy()
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{self.random_time}-1"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-
-        LOGGER.info("Step 4: Copy object from %s to other buckets in background", bucket_name)
-        args = {'s3_test_obj': s3_test_obj, 'bucket_name': bucket_name,
-                'object_name': self.object_name, 'bkt_obj_dict': bkt_obj_dict, 'output': output,
-                'file_path': self.multipart_obj_path, 'background': True, 'bkt_op': bkt_op,
-                'put_etag': put_etag}
-        thread = threading.Thread(target=self.ha_obj.create_bucket_copy_obj, args=(event,),
-                                  kwargs=args)
-        thread.daemon = True  # Daemonize thread
-        thread.start()
-        LOGGER.info("Step 4: Successfully started background process for copy object")
-        # While loop to sync this operation with background thread to achieve expected scenario
-        LOGGER.info("Waiting for creation of %s buckets", bkt_cnt)
-        bkt_list = list()
-        timeout = time.time() + 60 * 3
-        while len(bkt_list) < bkt_cnt:
-            time.sleep(HA_CFG["common_params"]["20sec_delay"])
-            bkt_list = s3_test_obj.bucket_list()[1]
-            if timeout < time.time():
-                LOGGER.error("Bucket creation is taking longer than 3 mins")
-                assert_utils.assert_true(False, "Please check background process logs")
-        time.sleep(HA_CFG["common_params"]["20sec_delay"])
-
-        LOGGER.info("Step 5: Starting pod again and checking cluster status")
-        event.set()
-        resp = self.ha_obj.restore_pod(pod_obj=self.node_master_list[0],
-                                       restore_method=self.restore_method,
-                                       restore_params={"deployment_name": self.deployment_name,
-                                                       "deployment_backup":
-                                                           self.deployment_backup,
-                                                       "num_replica": self.num_replica,
-                                                       "set_name": self.set_name},
-                                       clstr_status=True)
-        LOGGER.debug("Response: %s", resp)
-        assert_utils.assert_true(resp[0], f"Failed to restore pod by {self.restore_method} way")
-        LOGGER.info("Step 5: Successfully started the pod and cluster is online")
-        self.restore_pod = False
-        event.clear()
-
-        LOGGER.info("Step 6: Checking responses from background process")
-        thread.join()
-        responses = tuple()
-        while len(responses) < 3:
-            responses = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-
-        if not responses:
-            assert_utils.assert_true(False, "Background process failed to do copy object")
-
-        put_etag = responses[0]
-        exp_fail_bkt_obj_dict = responses[1]
-        failed_bkts = responses[2]
-        LOGGER.debug("Responses received from background process:\nput_etag: "
-                     "%s\nexp_fail_bkt_obj_dict: %s\nfailed_bkts: %s", put_etag,
-                     exp_fail_bkt_obj_dict, failed_bkts)
-        if len(exp_fail_bkt_obj_dict) == 0 and len(failed_bkts) == 0:
-            LOGGER.info("Copy object operation for all the buckets completed successfully. ")
-        elif failed_bkts or exp_fail_bkt_obj_dict:
-            assert_utils.assert_true(False, "Failed to do copy object when cluster was in degraded "
-                                            f"state. Failed buckets: \n{failed_bkts}"
-                                            f"\n{exp_fail_bkt_obj_dict}")
-        LOGGER.info("Step 6: Successfully completed copy object operation is background")
-
-        LOGGER.info("Step 7.1: Download the objects copied in healthy cluster and verify checksum")
-        for key, val in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=key, key=val)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in Etag verification of "
-                                                          f"object {val} of bucket {key}. Put and "
-                                                          f"Get Etag mismatch")
-        LOGGER.info("Step 7.1: Download the objects copied in step 4 and verify checksum")
-        for key, val in bkt_obj_dict1.items():
-            resp = s3_test_obj.get_object(bucket=key, key=val)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in Etag verification of "
-                                                          f"object {val} of bucket {key}. Put and "
-                                                          f"Get Etag mismatch")
-        LOGGER.info("Step 7: Successfully downloaded the object and verified the checksum")
-
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{self.random_time}-2"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-
-        LOGGER.info("Step 8: Perform copy object from %s bucket to other buckets verify copy object"
-                    " etags", bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  put_etag=put_etag,
-                                                  bkt_op=bkt_op)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        LOGGER.info("Step 8: Performed copy object from %s bucket to other buckets verified copy "
-                    "object etags", bucket_name)
-
-        LOGGER.info("Step 9: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 9: Downloaded copied objects & verify etags.")
-
-        LOGGER.info("ENDED: Test to verify copy object during data pod restart")
 
     @pytest.mark.ha
     @pytest.mark.lc
@@ -1960,7 +1108,6 @@ class TestDataPodRestart:
         LOGGER.info("Step 2: Successfully shutdown data pod %s. Verified cluster and "
                     "services states are as expected & remaining pods status is online.", pod_name)
         self.restore_pod = True
-        output = Queue()
         event = threading.Event()  # Event to be used to send intimation of pod restart
         LOGGER.info("Step 3: Start READ/WRITEs/VERIFY with variable object sizes in background")
         if CMN_CFG["dtm0_disabled"]:
